requirements.txt
feedparser==6.0.11
requests==2.31.0
anthropic==0.39.0
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.10.3
